# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My-Trip API Documentation",
        "description": "API documentation for My-Trip Platform",
        "contact": {
            "email": "support@mytrip.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/wjhtak-site/public",
            "description": "Development Server"
        },
        {
            "url": "https://blueviolet-hummingbird-437500.hostingersite.com/public",
            "description": "Production Server 1"
        },
        {
            "url": "https://lavenderblush-bear-243464.hostingersite.com/public/",
            "description": "Production Server 2"
        }
    ],
    "paths": {
...
        "/api/payment/initiate": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Initiate HyperPay checkout",
                "description": "Initializes a payment session and returns a Checkout ID for the mobile SDK.",
                "operationId": "initiatePayment",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "payment_type"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "enum": [
                                            "mada",
                                            "visa_master",
                                            "apple_pay"
                                        ],
                                        "example": "visa_master"
                                    },
                                    "order_id": {
                                        "type": "string",
                                        "example": "ORDER-12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checkout initialized successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Checkout initialized successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "8AC7A4CA728C..."
                                                },
                                                "result": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Payment Gateway Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment/verify": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Verify HyperPay payment status",
                "description": "Checks the status of a payment using the Checkout ID after the transaction is completed on the client side.",
                "operationId": "verifyPayment",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "The language of the response (ar, en)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "checkout_id",
                                    "payment_type"
                                ],
                                "properties": {
                                    "checkout_id": {
                                        "type": "string",
                                        "example": "8AC7A4CA728C..."
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "enum": [
                                            "mada",
                                            "visa_master",
                                            "apple_pay"
                                        ],
                                        "example": "visa_master"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment status retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment successful."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Payment Failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}